+++ addOns/network/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Fixed
+- Do not close the client connection when the server closes it, if not required, to keep the client connection in good state and be used longer.
 
 ## [0.8.0] - 2023-05-03
 ### Added